Improve error handling in ubuntu install

+++ zsh/install.ubuntu.sh
@@ -1,7 +1,22 @@
 . "$DOTFILES/scripts/core/main.sh"
 
-apt install zsh
-log::result $? "Install Zsh"
+# Check version exisiting
+if ! platform::command_exists "zsh"
+then
+  apt install zsh
+  log::result $? "Set Zsh as default shell"
+fi
 
-chsh -s $(which zsh)
-log::result $? "Set Zsh as default shell"
+# Set as default shell
+ZSH_SHELL_LOC=$(which zsh)
+if [ "$SHELL" != "$ZSH_SHELL_LOC" ]
+then
+  chsh -s /bin/zsh
+  log::result $? "Set Zsh as default shell"
+fi
+
+# Install zplug for the next bit
+if [[ -z $ZPLUG_HOME ]]; then
+  export ZPLUG_HOME=~/.zplug
+  git clone https://github.com/zplug/zplug $ZPLUG_HOME
+fi
